refactor: Validation Init 타입 추가
